feat: Add user account deletion and admin account suspension functionality
This commit adds the ability for users to delete their own accounts and for admins to suspend and unsuspend other user accounts. It also includes error handling to prevent server crashes and improvements to the book deletion process, such as checking for the existence of the book folder before removing it. Additionally, if a user logs in after disabling their account, the scheduled deletion is canceled and they receive an email notification about the account reinstatement. Security protocols have been implemented in certain routes, and the delete book by admin functionality now includes a reason for deletion. The swagger documentation has also been updated to allow book deletion and the associated image folder is now removed when a book is deleted.